Convert to machine sized integers
If we were to use u64/i64 across the board, we'd silently overflow and
get incorrect results for some inputs on systems with 32-bit long.

If we were to use u32/i32 across the board, we'd lose functionality on
systems with 64-bit long.

NumPy on x86_64:

    >>> numpy.random.binomial(0x80000000, 0.3)
    644264428

NumPy on i686:

    >>> numpy.random.binomial(0x80000000, 0.3)
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
        File "mtrand.pyx", line 3514, in mtrand.RandomState.binomial
        (numpy/random/mtrand/mtrand.c:17090)
        OverflowError: Python int too large to convert to C long

On every system I'm aware of, sizeof(intptr_t) == sizeof(long), so
Rust's int and c_long should be the same (similar for uint). So, for
mimicing the behavior of NumPy, this should be a safe bet.

It may be worth thinking about the distributions and restricting their
input on a case-by-case basis. For example, observe the (poor?) behavior
of NumPy:

    >>> np.random.geometric(0.00000000000000005)
    -9223372036854775808  # == -0x8000000000000000